Removed setup command from waf build context.
setup command is responsible for installing all missing dependencies, and
configure command, that creates build context responsible for checking if all
dependencies are satisfied. So we can't use setup komand if build context is
created and we can't use configure command since required dependencies are not
installed, which is setup komand responsibility.

Need to find out better way, how to mand dependencies configurable and make it
work together with configure command.