Sync CouchDB views for scrapy
Currently all CouchDB views are synchorinzed using couchdbkit, but this works
only for Django apps and single app can synchronize single database.

manoseimas/scrapy is not Django app and it has to deal with both, sittings and
nodes databases, o couchdbkit can not be used here.

To sync views for scrapy I created simple script initscrapy, to do the job.