Fix lobbyist declaration scrapers in MySQL STRICT mode
I managed to misread the Django documentation into assuming that if I
declare a model field as

  CharField(null=False, blank=True)

then Django will store None values as ''.

That's not the case.

I didn't notice, because by default MySQL happily accepts SQL of the form

  UPDATE lobbyists_lobbyist SET status = NULL WHERE id = 123;

even when the column is explicitly set to NOT NULL.

Now, inside a Vagrant environment, we have manoseimas/settings/vagrant.py
that sets the MySQL connection init string to something that includes

  SET GLOBAL sql_mode=STRICT_ALL_TABLES;

which tells MySQL to wake up and pay attention.

Since I once ran `make` outside Vagrant, I had a buildout.cfg that did
_not_ point to manoseimas.settings.vagrant.