Whip state actor classifier into shape
I replaced the testcases with the desired classification of all of the
current data set, provided by TILS, and then whacked the code to get it
to run.

I'm not happy about the "known_state_actors" global variable, but it'll
do as a short-term solution.

A long-term solution would store the "state actor?" bit in the database,
and would cross-reference against the MPs table when inserting new
entries, and would provide the possibility for admins to override the
classification guess.

Fixes #101.