Fix annoying database issue when running tests
Now database name is taken from 'read_default_file' parameter which points to a
my.cnf configuration, where database name should be defined.

But this does not work when running tests for a single file, database name is
set to 'test_', but when running all tests, then database name is set to
'test_manoseimas'.

Setting database name in settings explicitly fixes this.